KOGITO-6396 - Non unique test case IDs in process-kafka-* examples
