+++ process-kafka-quickstart-quarkus/src/test/java/org/acme/travel/tests/messaging/quarkus/MessagingIT.java
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.acme.travel;
+package org.acme.travel.tests.messaging.quarkus;
 
 import java.net.URI;
 import java.time.OffsetDateTime;
 import java.util.Optional;
-import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
@@ -26,6 +25,7 @@ import java.util.stream.IntStream;
 
 import javax.inject.Inject;
 
+import org.acme.travel.Traveller;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -51,13 +51,11 @@ import static org.junit.jupiter.api.Assertions.fail;
 
 @QuarkusTest
 @QuarkusTestResource(KafkaQuarkusTestResource.class)
-public class MultiMessagingIT {
+public class MessagingIT {
 
     public static final String TOPIC_PRODUCER = "travellers";
-    public static final String TOPIC_PROCESSED_CONSUMER = "processedtravellers";
-    public static final String TOPIC_CANCEL_CONSUMER = "cancelledtravellers";
-
-    private static Logger LOGGER = LoggerFactory.getLogger(MultiMessagingIT.class);
+    public static final String TOPIC_CONSUMER = "processedtravellers";
+    private static Logger LOGGER = LoggerFactory.getLogger(MessagingIT.class);
 
     @Inject
     private ObjectMapper objectMapper;
@@ -72,6 +70,13 @@ public class MultiMessagingIT {
         kafkaClient = new KafkaTestClient(kafkaBootstrapServers);
     }
 
+    @AfterEach
+    public void close() {
+        if (kafkaClient != null) {
+            kafkaClient.shutdown();
+        }
+    }
+
     @Test
     public void testProcess() throws InterruptedException {
         objectMapper.configure(JsonParser.Feature.ALLOW_UNQUOTED_FIELD_NAMES, true);
@@ -80,15 +85,16 @@ public class MultiMessagingIT {
         final int count = 3;
         final CountDownLatch countDownLatch = new CountDownLatch(count);
 
-        kafkaClient.consume(Set.of(TOPIC_PROCESSED_CONSUMER, TOPIC_CANCEL_CONSUMER), s -> {
+        kafkaClient.consume(TOPIC_CONSUMER, s -> {
             LOGGER.info("Received from kafka: {}", s);
             try {
                 JsonNode event = objectMapper.readValue(s, JsonNode.class);
                 Traveller traveller = objectMapper.readValue(event.get("data").toString(), Traveller.class);
-                assertEquals(!traveller.getNationality().equals("American"), traveller.isProcessed());
+                assertTrue(traveller.isProcessed());
                 assertTrue(traveller.getFirstName().matches("Name[0-9]+"));
                 assertTrue(traveller.getLastName().matches("LastName[0-9]+"));
                 assertTrue(traveller.getEmail().matches("email[0-9]+"));
+                assertTrue(traveller.getNationality().matches("Nationality[0-9]+"));
                 countDownLatch.countDown();
             } catch (JsonProcessingException e) {
                 LOGGER.error("Error parsing {}", s, e);
@@ -97,15 +103,11 @@ public class MultiMessagingIT {
         });
 
         IntStream.range(0, count)
-                .mapToObj(i -> new Traveller("Name" + i, "LastName" + i, "email" + i, getNationality(i)))
+                .mapToObj(i -> new Traveller("Name" + i, "LastName" + i, "email" + i, "Nationality" + i))
                 .forEach(traveller -> kafkaClient.produce(generateCloudEvent(traveller), TOPIC_PRODUCER));
 
         countDownLatch.await(10, TimeUnit.SECONDS);
         assertEquals(0, countDownLatch.getCount());
-    }
-
-    private String getNationality(int i) {
-        return i % 2 == 0 ? "American" : "Spanish";
     }
 
     private String generateCloudEvent(Traveller traveller) {
