+++ process-kafka-quickstart-springboot/src/test/java/org/acme/travel/tests/messaging/springboot/MessagingIT.java
@@ -13,22 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.acme.travel;
+package org.acme.travel.tests.messaging.springboot;
 
 import java.net.URI;
 import java.time.OffsetDateTime;
-import java.util.Arrays;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.IntStream;
 
+import org.acme.travel.Traveller;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.kie.kogito.test.springboot.kafka.KafkaTestClient;
 import org.kie.kogito.testcontainers.springboot.KafkaSpringBootTestResource;
-import org.kie.kogito.tests.KogitoKafkaMultiSpringbootApplication;
+import org.kie.kogito.tests.KogitoKafkaQuickstartSpringbootApplication;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -47,15 +47,13 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
-@SpringBootTest(classes = KogitoKafkaMultiSpringbootApplication.class)
+@SpringBootTest(classes = KogitoKafkaQuickstartSpringbootApplication.class)
 @ContextConfiguration(initializers = KafkaSpringBootTestResource.class)
-public class MultiMessagingIT {
+public class MessagingIT {
 
     public static final String TOPIC_PRODUCER = "travellers";
-    public static final String TOPIC_PROCESSED_CONSUMER = "processedtravellers";
-    public static final String TOPIC_CANCEL_CONSUMER = "cancelledtravellers";
-
-    private static Logger LOGGER = LoggerFactory.getLogger(MultiMessagingIT.class);
+    public static final String TOPIC_CONSUMER = "processedtravellers";
+    private static Logger LOGGER = LoggerFactory.getLogger(MessagingIT.class);
 
     @Autowired
     private ObjectMapper objectMapper;
@@ -71,15 +69,16 @@ public class MultiMessagingIT {
         final int count = 3;
         final CountDownLatch countDownLatch = new CountDownLatch(count);
 
-        kafkaClient.consume(Arrays.asList(TOPIC_PROCESSED_CONSUMER, TOPIC_CANCEL_CONSUMER), s -> {
+        kafkaClient.consume(TOPIC_CONSUMER, s -> {
             LOGGER.info("Received from kafka: {}", s);
             try {
                 JsonNode event = objectMapper.readValue(s, JsonNode.class);
                 Traveller traveller = objectMapper.readValue(event.get("data").toString(), Traveller.class);
-                assertEquals(!traveller.getNationality().equals("American"), traveller.isProcessed());
+                assertTrue(traveller.isProcessed());
                 assertTrue(traveller.getFirstName().matches("Name[0-9]+"));
                 assertTrue(traveller.getLastName().matches("LastName[0-9]+"));
                 assertTrue(traveller.getEmail().matches("email[0-9]+"));
+                assertTrue(traveller.getNationality().matches("Nationality[0-9]+"));
                 countDownLatch.countDown();
             } catch (JsonProcessingException e) {
                 LOGGER.error("Error parsing {}", s, e);
@@ -88,15 +87,11 @@ public class MultiMessagingIT {
         });
 
         IntStream.range(0, count)
-                .mapToObj(i -> new Traveller("Name" + i, "LastName" + i, "email" + i, getNationality(i)))
+                .mapToObj(i -> new Traveller("Name" + i, "LastName" + i, "email" + i, "Nationality" + i))
                 .forEach(traveller -> kafkaClient.produce(generateCloudEvent(traveller), TOPIC_PRODUCER));
 
         countDownLatch.await(10, TimeUnit.SECONDS);
         assertEquals(0, countDownLatch.getCount());
-    }
-
-    private String getNationality(int i) {
-        return i % 2 == 0 ? "American" : "Spanish";
     }
 
     private String generateCloudEvent(Traveller traveller) {
